backend/requirements.txt
google-adk==0.1.0
requests==2.31.0
python-dotenv==1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
pandas==2.1.1
nltk==3.8.1
pytest==7.4.3
yfinance==0.2.28
yahoo_fin==0.8.9.1
